create a hollow cylinder for the gantry path, just do gantry head

+++ DosCheckRun.py
@@ -2,13 +2,13 @@ from connect import *
 import math
 
 
-def make_cylinder(case, examination, couch_angle, iso, name, color='red', radius=42, length=80):
+def make_cylinder(case, examination, couch_angle, iso, roi_names, name, color='red', radius=42, length=80):
     px, py, pz = iso
     cosCouch = math.cos(math.pi * couch_angle / 180)
     sinCouch = math.sin(math.pi * couch_angle / 180)
-    try:
+    if name not in roi_names:
         retval_0 = case.PatientModel.CreateRoi(Name=name, Color=color, Type="Undefined")
-    except:
+    else:
         case.PatientModel.RegionsOfInterest[name].DeleteRoi()
         retval_0 = case.PatientModel.CreateRoi(Name=name, Color=color, Type="Undefined")
 
@@ -16,15 +16,36 @@ def make_cylinder(case, examination, couch_angle, iso, name, color='red', radius
                                     Examination=examination, Center={'x': px, 'y': py, 'z': pz},
                                     Representation="Voxels", VoxelSize=None)
 
+    retval_0.CreateAlgebraGeometry(Examination=examination, Algorithm="Auto",
+                                  ExpressionA={'Operation': "Union", 'SourceRoiNames': [name],
+                                               'MarginSettings': {'Type': "Expand", 'Superior': 0,
+                                                                  'Inferior': 0, 'Anterior': 5, 'Posterior': 5,
+                                                                  'Right': 5, 'Left': 5}},
+                                  ExpressionB={'Operation': "Union", 'SourceRoiNames': [name],
+                                               'MarginSettings': {'Type': "Expand", 'Superior': 1,
+                                                                  'Inferior': 1, 'Anterior': 0, 'Posterior': 0,
+                                                                  'Right': 0, 'Left': 0}},
+                                  ResultOperation="Subtraction",
+                                  ResultMarginSettings={'Type': "Expand", 'Superior': 0, 'Inferior': 0,
+                                                        'Anterior': 0, 'Posterior': 0, 'Right': 0, 'Left': 0})
     # rotate the cylinder
     transmat = {'M11': cosCouch, 'M12': 0, 'M13': sinCouch, 'M14': 1 + px - px * cosCouch - pz * sinCouch,
                 'M21': 0, 'M22': 1, 'M23': 0, 'M24': 1, 'M31': -1 * sinCouch, 'M32': 0, 'M33': cosCouch,
                 'M34': 1 + pz - pz * cosCouch + px * sinCouch, 'M41': 0, 'M42': 0, 'M43': 0, 'M44': 1}
     retval_0.TransformROI3D(Examination=examination, TransformationMatrix=transmat)
 
+def make_box(case, examination, iso, roi_names, name, color='red', lr=70, sup_inf=25, ap=50):
+    px, py, pz = iso
+    if name not in roi_names:
+        retval_0 = case.PatientModel.CreateRoi(Name=name, Color=color, Type="Undefined")
+        retval_0.CreateBoxGeometry(Size={'x': lr, 'y': ap, 'z': sup_inf},
+                                   Examination=examination, Center={'x': px, 'y': py, 'z': pz},
+                                   Representation="Voxels", VoxelSize=None)
+
 
 def run(case, beam_set, examination):
     couch_isocenters = {}
+    roi_names = [i.Name for i in case.PatientModel.RegionsOfInterest]
     for beam in beam_set.Beams:
         beam_name = beam.Name
         couch_angle = beam.CouchRotationAngle
@@ -40,20 +61,39 @@ def run(case, beam_set, examination):
                 run_check = True
             if not run_check:
                 continue
-            make_cylinder(case, examination, couch_angle, iso, "GantryHead_42cm", color='red', radius=42,
-                          length=80)
-            overlap_name = beam_name + '_' + 'GantryHead'
-            try:
-                retval_0 = case.PatientModel.CreateRoi(Name=overlap_name, Color='blue', Type="Undefined")
-            except:
-                case.PatientModel.RegionsOfInterest[overlap_name].DeleteRoi()
-                retval_0 = case.PatientModel.CreateRoi(Name=overlap_name, Color='blue', Type="Undefined")
-
-            make_cylinder(case, examination, couch_angle, iso, "kVpanel_51cm", color='yellow', radius=51,
-                          length=80)
-
-
-    poiList = case.PatientModel.StructureSets[examination.Name].PoiGeometries
+            """
+            If all the beams share the same iso and couch, only have to run once
+            """
+            for roi_name, radius in zip(["GantryHead", "kVPanel"], [25, 51]):
+                make_cylinder(case, examination, couch_angle, iso, roi_names, roi_name, color='red',
+                              radius=radius, length=25)
+                overlap_name = "Potential_Collision" + '_' + beam_name
+                if overlap_name not in roi_names:
+                    ret_val = case.PatientModel.CreateRoi(Name=overlap_name, Color='blue', Type="Undefined")
+                else:
+                    case.PatientModel.RegionsOfInterest[overlap_name].DeleteRoi()
+                    ret_val = case.PatientModel.CreateRoi(Name=overlap_name, Color='blue', Type="Undefined")
+                ret_val.CreateAlgebraGeometry(Examination=examination, Algorithm="Auto",
+                                              ExpressionA={'Operation': "Union", 'SourceRoiNames': [r"External ROI"],
+                                                           'MarginSettings': {'Type': "Expand", 'Superior': 5,
+                                                                              'Inferior': 5, 'Anterior': 5, 'Posterior': 0,
+                                                                              'Right': 5, 'Left': 5}},
+                                              ExpressionB={'Operation': "Union", 'SourceRoiNames': [roi_name],
+                                                           'MarginSettings': {'Type': "Expand", 'Superior': 0,
+                                                                              'Inferior': 0, 'Anterior': 0, 'Posterior': 0,
+                                                                              'Right': 0, 'Left': 0}},
+                                              ResultOperation="Intersection",
+                                              ResultMarginSettings={'Type': "Expand", 'Superior': 0, 'Inferior': 0,
+                                                                    'Anterior': 0, 'Posterior': 0, 'Right': 0, 'Left': 0})
+                if case.PatientModel.StructureSets[examination.Name].RoiGeometries[overlap_name].HasContours():
+                    print("Potential collision!" + ' ' + overlap_name)
+                else:
+                    case.PatientModel.RegionsOfInterest[overlap_name].DeleteRoi()
+                case.PatientModel.RegionsOfInterest[roi_name].DeleteRoi()
+                """
+                Why check gantry and kv-panels? If the gantry passes, so will the kv-panels
+                """
+                break
 
 
 def main():
